feat: Enviando funcionalidade de exclusão, adição e edição

+++ FeatureJuan/Repository/TimeRepository.cs
@@ -37,10 +37,35 @@ namespace FeatureJuan.Repository
             await _context.SaveChangesAsync();
         }
 
-        public async Task DeleteTime(int id)
+        public async Task DeleteTime(int EquipeId)
         {
-            var equipe = await _context.Equipes.FindAsync(id);
+            var equipe = await _context.Equipes.FindAsync(EquipeId);
+            if (equipe != null)
+            {
             _context.Equipes.Remove(equipe);
+            await _context.SaveChangesAsync();
+            }
+        }
+
+        public async Task<Equipe> GetTimeById(int EquipeId)
+        {
+            return await _context.Equipes
+                .Select(e => new Equipe
+                {
+                    EquipeId = e.EquipeId,
+                    Nome = e.Nome,
+                    QuantidadeIntegrantes = e.QuantidadeIntegrantes,
+                    Divisao = new Divisao
+                    {
+                        DivisaoId = e.Divisao.DivisaoId,
+                        Nome = e.Divisao.Nome
+                    }
+                })
+                .FirstOrDefaultAsync(x => x.EquipeId == EquipeId);
+        }
+        public async Task EditarTime(Equipe equipe)
+        {
+            _context.Equipes.Update(equipe);
             await _context.SaveChangesAsync();
         }
     }
